# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Public Works Transparency Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "identity and reference"
        ],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/v1/auth/otp/request": {
      "post": {
        "tags": [
          "identity and reference"
        ],
        "summary": "Request Otp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP requested"
          }
        }
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "identity and reference"
        ],
        "summary": "Verify Otp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "identity and reference"
        ],
        "summary": "Me",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginBody": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "PhoneBody": {
        "type": "object",
        "required": [
          "phone"
        ],
        "properties": {
          "phone": {
            "type": "string"
          }
        }
      },
      "OtpBody": {
        "type": "object",
        "required": [
          "phone",
          "otp"
        ],
        "properties": {
          "phone": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          }
        }
      }
    }
  }
}